phone user delete

+++ src/edu/sjsu/cmpe275/lab2/dao/CreatePhone.java
@@ -93,9 +93,11 @@ public class CreatePhone {
 	    entitymanager.getTransaction( ).begin( );
 	      Phone phone = new Phone();
 		  phone = entitymanager.find(Phone.class, phoneId);
+		 if(phone.getUser().size()== 0){
 		  entitymanager.remove(phone);
 		  entitymanager.getTransaction().commit();
 		  entitymanager.close();
+		 }
 	}
